% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBMI.R
\name{calcBMI}
\alias{calcBMI}
\title{Calculate BMI}
\usage{
calcBMI(height, weight, age = rep(NA, length = length(height)))
}
\arguments{
\item{height}{Height in meters}

\item{weight}{Weight in kilograms}

\item{age}{Age in months}
}
\value{
A vector with BMI in kg/m^2 or weight-for-age in kg.
}
\description{
\code{height} and \code{weight} are required arguments. If \code{age} is supplied,
then weight-for-age will be calculated for ages less than 24 months.
}
\examples{
data(nhanes)
with(nhanes, calcBMI(height, weight, age))
}

