% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZScore.R
\name{calcZScore}
\alias{calcZScore}
\title{Calculate Z-score of BMI or weight-for-age}
\usage{
calcZScore(age, sex, bmi)
}
\arguments{
\item{age}{Age in months. Must be integer-values from 0 to 240.}

\item{sex}{Sex. Must be "Male"/"Female"}

\item{bmi}{BMI in kg/m^2 or weight-for-age in kg.}
}
\value{
A vector of Z-Scores.
}
\description{
Calculate Z-score of BMI or weight-for-age
}
\examples{
data(nhanes)
nhanes$bmi <- with(nhanes, calcBMI(height, weight, age))
with(nhanes, calcZScore(age, sex, bmi))
}
\seealso{
\code{\link{chart}} for information about how Z-Scores and percentiles are calculated.
}

