% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValid.R
\name{isValid}
\alias{isValid}
\title{Validate data}
\usage{
isValid(age, bmi, zscore)
}
\arguments{
\item{age}{Age in months}

\item{bmi}{BMI in kg/m^2 or weight-for-age in kg}

\item{zscore}{Z-Score of BMI or weight-for-age}
}
\value{
A logical vector that answers the question, are the data validated?
}
\description{
The function returns `FALSE` when:
}
\details{
\itemize{
  \item Age is 24 months or greater and BMI is less than 6 or greater than 100
  \item Age is between 24 and 240 months inclusively, and Z-Score is less than -4 or greater than 5
  \item Age is less than 24 months and Z-Score is less than -6 or greater than 5
}
}
\examples{
data(nhanes)
nhanes$bmi <- with(nhanes, calcBMI(height, weight, age))
nhanes$zscore <- with(nhanes, calcZScore(age, sex, bmi))
nhanes$valid <- with(nhanes, isValid(age, bmi, zscore))
table(nhanes$valid)
nhanes[!nhanes$valid,]
}

