% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{NHANES 2013-2014 Data}
\format{\describe{
 \item{seqn}{Respondent sequence number}
 \item{age}{Age in months}
 \item{sex}{Sex, "Male"/"Female"}
 \item{height}{Height in meters. Missing for infants.}
 \item{weight}{Weight in kilograms}
}}
\usage{
nhanes
}
\description{
This data.frame contains a subset of information from
\href{http://wwwn.cdc.gov/nchs/nhanes/search/nhanes13_14.aspx}{NHANES 2013-2014}
Demographics and Examination data.
}
\keyword{datasets}

