% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValid.R
\name{isValid}
\alias{isValid}
\title{Validate weight measurements by common protocols}
\usage{
isValid(height, weight, age, sex)
}
\arguments{
\item{height}{Height in meters}

\item{weight}{Weight in kilograms}

\item{age}{Age in months}

\item{sex}{Sex}
}
\value{
A logical vectore
}
\description{
Validate weight measurements by common protocols
}
\examples{
data(nhanes)
with(nhanes, calcBMI(height, weight, age))
}

